+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-This project adheres to [Semantic Versioning](http://semver.org/).
+## UNRELEASED
+
+- Bug: Move plugin types and update to webpack v5. [#324](https://github.com/FormidableLabs/webpack-dashboard/issues/324)
 
 ## 3.3.6